requirements.txt
pypdf
weaviate-client
nltk
dotenv